Bunch of things
Changed the lighting model to subsurface so our foliage is prettier

Added stuff to the master material for the sub surface scattering and
also added a constant to control a materials roughness

Added a new grass texture

Made a landscape material with two layers for now.

Added all the foliage stuff to the foliage brush and gave most of them
some settings that should work.

Fixed the distance drop material function so it works now.

Grass gets culled now.