Added colour changing to master material
Just tick the static switch for colour shifting and enable and change
the colour shift vector in order to crudely change the colours of
textures.  This is so if anyone wants to see how anything might look in
a different colour they can quickly test it out in engine by changing
material instance settings.  Ideally changes will ultimately made to the
texture, but this is a quick and dirty way to see how things might look.