Tree and Web
Added a breakable tree and a non breakable web (not sure how to do it
just yet)

I also made a new material for the web.  Since it needs translucency I
figure it'd be better to just make it its own custom material than to
change the master to betranslucent. The web material also has some depth
nonsense on the alpha so it doesn't look like it clips through
everything.  WOOO